% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{add_parentNodes}
\alias{add_parentNodes}
\title{Add the parent nodes to a vector of node.}
\usage{
add_parentNodes(vec.node, graph = graph)
}
\arguments{
\item{vec.node}{An vector of node names.}

\item{graph}{The igraph object of GIN.}
}
\value{
A vector of node names.
}
\description{
Find the intermediate nodes, and add the substrate/enzyme nodes to the input vector.
}
