% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{add_prcFormation_and_GRN}
\alias{add_prcFormation_and_GRN}
\title{Add complex formation paths and gene regulatory network to an igraph object}
\usage{
add_prcFormation_and_GRN(g.subgraph, input.gs, graph.grn, cdglist)
}
\arguments{
\item{g.subgraph}{The igraph object which needs to add complex formation paths and GRN.}

\item{input.gs}{A vector of input genes.}

\item{graph.grn}{An igraph object created from edgelist of Gene Regulatory Network.}

\item{cdglist}{A vector of all coding genes' gene symbols.}
}
\value{
An igraph object which contains complex formation paths and gene regulatory networks.
}
\description{
When inducing a sub-graph from GIN based on an input gene/chemical list, lots of protein complexes are included. However, the paths that form these complexes are not generated, since their protein members may not be in the input gene/chemical list. This function finds all the protein complexes and generate paths for the formation of the complexes. The paths may be sufficient for simulation of signal flows, but may not be accurate, therefore needs manual curations. This function will also add gene regulatory networks to the input igraph object.
}
