% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{adjustNode_color_size}
\alias{adjustNode_color_size}
\title{Adjust the layout of nodes in a subgraph}
\usage{
adjustNode_color_size(
  g,
  input.gs,
  node.color = c("yellow", "brown", "grey", "blue", "orange", "red"),
  node.size = c(5, 3, 7),
  cdglist.this = cdglist
)
}
\arguments{
\item{g}{An igraph object. Should be a subgraph induced from GIN based on input.gs.}

\item{input.gs}{A list of genes/chemicals.l}

\item{node.color}{A vector of colors for: chemicals, complexes, intermediates, promoter regions, RNA transcripts, and proteins. Default is "yellow","brown","grey","blue","orange","red".}

\item{node.size}{A numeric vector of three elements to set the size of different types of nodes: the nodes not in the input list; the intermediate nodes; the nodes in the input list. Default is 5,3,7}

\item{cdglist.this}{A list of coding genes' symbols.}
}
\value{
An igraph object adjusted for nodes' colors and sizes.
}
\description{
Adjust the color and size of the nodes.
}
