% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{analyze_node_composition}
\alias{analyze_node_composition}
\title{Analyze the composition of nodes' types}
\usage{
analyze_node_composition(g.sub, cdglist = cdglist)
}
\arguments{
\item{g.sub}{An igraph object. The names of the nodes will be fetched by "V(g.sub)$name".}

\item{cdglist}{A list of coding genes' gene symbols.}
}
\value{
A named numeric vector which contains the counts of different types of nodes.
}
\description{
Count the number of nodes in different types.
}
