% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{basic_node}
\alias{basic_node}
\title{Generate the text of a node.}
\usage{
basic_node(
  id,
  height.geo = "20.0",
  width.geo = "20.0",
  x.geo = runif(n = 1, min = 200, max = 800),
  y.geo = runif(n = 1, min = 200, max = 800),
  color.fill = "#00CC33",
  transparent = "false",
  color.border = "#000000",
  type.border = "line",
  width.border = "1.0",
  alignment.nl = "center",
  autoSizePolicy.nl = "node_width",
  fontFamily.nl = "Arial Black",
  fontSize.nl = "13",
  fontStyle.nl = "plain",
  hasBackgroundColor.nl = "false",
  hasLineColor.nl = "false",
  height.nl = "22.33203125",
  horizontalTextPosition.nl = "center",
  iconTextGap.nl = "4",
  modelName.nl = "internal",
  modelPosition.nl = "c",
  textColor.nl = "#000000",
  verticalTextPosition.nl = "bottom",
  visible.nl = "true",
  width.nl = "20.0",
  x.nl = "0",
  xmlspace.nl = "preserve",
  y.nl = "-1.166015625",
  text.show,
  shape = "ellipse"
)
}
\arguments{
\item{id}{The id of a node. The format of this id is "n" plus a number, for example "n1", "n2"...}

\item{height.geo}{Character. The "height" parameter in "Geometry" section. Default is "20.0".}

\item{width.geo}{Character. The "width" parameter in "Geometry" section. Default is "20.0".}

\item{x.geo}{Character. The "x" parameter in "Geometry" section. Default is randomly generated by runif(n=1,min=200, max =800).}

\item{y.geo}{Character. The "y" parameter in "Geometry" section. Default is randomly generated by runif(n=1,min=200, max =800).}

\item{color.fill}{Character. The "color" parameter in "Fill" section. Default is "#00CC33".}

\item{transparent}{Character. The "transparent" parameter in "Fill" section. Default is "false".}

\item{color.border}{Character. The "color" parameter in "BorderStyle" section. Default is "#000000".}

\item{type.border}{Character. The "type" parameter in "BorderStyle" section. Default is "line".}

\item{width.border}{Character. The "width" parameter in "BorderStyle" section. Default is "1.0".}

\item{alignment.nl}{Character. The "alignment" parameter in "NodeLabel" section. Default is "center".}

\item{autoSizePolicy.nl}{Character. The "autoSizePolicy" parameter in "NodeLabel" section. Default is "node_width".}

\item{fontFamily.nl}{Character. The "fontFamily" parameter in "NodeLabel" section. Default is "Arial Black".}

\item{fontSize.nl}{Character. The "fontSize" parameter in "NodeLabel" section. Default is "13".}

\item{fontStyle.nl}{Character. The height parameter in "NodeLabel" section. Default is "plain".}

\item{hasBackgroundColor.nl}{Character. The "hasBackgroundColor" parameter in "NodeLabel" section. Default is "false".}

\item{hasLineColor.nl}{Character. The "hasLineColor" parameter in "NodeLabel" section. Default is "false".}

\item{height.nl}{Character. The "height" parameter in "NodeLabel" section. Default is "22.33203125".}

\item{horizontalTextPosition.nl}{Character. The "horizontalTextPosition" parameter in "NodeLabel" section. Default is "center".}

\item{iconTextGap.nl}{Character. The "iconTextGap" parameter in "NodeLabel" section. Default is "4".}

\item{modelName.nl}{Character. The "modelName" parameter in "NodeLabel" section. Default is "internal".}

\item{modelPosition.nl}{Character. The "modelPosition" parameter in "NodeLabel" section. Default is "c".}

\item{textColor.nl}{Character. The "textColor" parameter in "NodeLabel" section. Default is "#000000".}

\item{verticalTextPosition.nl}{Character. The "verticalTextPosition" parameter in "NodeLabel" section. Default is "bottom".}

\item{visible.nl}{Character. The "visible" parameter in "NodeLabel" section. Default is "true".}

\item{width.nl}{Character. The "width" parameter in "NodeLabel" section. Default is "20.0".}

\item{x.nl}{Character. The "x" parameter in "NodeLabel" section. Default is "0".}

\item{xmlspace.nl}{Character. The "xmlspace" parameter in "NodeLabel" section. Default is "preserve".}

\item{y.nl}{Character. The "y" parameter in "NodeLabel" section. Default is "-1.166015625".}

\item{text.show}{Character. The text to show on the node when imported into an editor.}

\item{shape}{Character. The "type" parameter in "Shape" section. Default is "ellipse".}
}
\value{
A character string of the texts for a node.
}
\description{
The basic function to generate the texts for a node. The default texts are based upon the schema of a transition node.
}
