% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{convert_complexVec_to_matrix}
\alias{convert_complexVec_to_matrix}
\title{Convert a vector of complex into a matrix}
\usage{
convert_complexVec_to_matrix(complexVec)

convert_complexVec_to_matrix(complexVec)
}
\arguments{
\item{complexVec}{An vector of complex names. The protein members are separated by "_".}
}
\value{
A two-dimensional matrix. The rows are complexes, the columns are members. If a complex contains certain members, then the corresponding elements in the matrix are set to 1, otherwise set to 0.

A numeric matrix of 0 and 1. The rows are protein complexes. The columns are protein members. 1 means the protein member is one of the components of the complex. 0 means not.
}
\description{
Complexes are groups of members, which may be shared by different complexes. This function convert a given vector of complexes into a matrix to describe the members. It can be used for calculation of the intersections between complexes.

Convert a vector of protein complex into a matrix.
}
