% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{convert_edge_to_graphml}
\alias{convert_edge_to_graphml}
\title{Convert an edge into a list}
\usage{
convert_edge_to_graphml(
  id.edge.num,
  source,
  target,
  edgeType = c("positive", "negative"),
  graphml
)
}
\arguments{
\item{id.edge.num}{The number of an edge id. An edge id consists of "e" and a number, for example "e1", "e2"...}

\item{source}{The source node id.}

\item{target}{The target node id.}

\item{edgeType}{The type of the edge. Should be either "positive" or "negative".}

\item{graphml}{The GraphML list.}
}
\value{
A list of two elements. The first one is the text of an edge, named "edge_texts". The second is "metainfo", which is a dataframe with three columns, id, textShow (stores the source and target), and type.
}
\description{
Generate a list of edge text and the meta info. Meta info is useful for accelerating downstream conversion.
}
