% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{convert_edgelist_to_graphml}
\alias{convert_edgelist_to_graphml}
\title{Convert an edge list into a list}
\usage{
convert_edgelist_to_graphml(el, graphml, metainfo.node, cdglist)
}
\arguments{
\item{el}{An edge list with three columns, the text.show of the source node, the text.show of the target node, and the edge types.}

\item{graphml}{The GraphML list.}

\item{metainfo.node}{A dataframe containing the meta info of nodes. This dataframe consists of three columns, id, textShow, and type}

\item{cdglist}{A list of the coding genes' gene symbols.}
}
\value{
A list of two elements. The first one is the list of nodes and edges, named "graphml". The second is "metainfo", which is a dataframe with three columns, id, textShow (stores the source and target), and type.
}
\description{
Generate a list of edge texts and the meta info according to an input edgelist. The edgelist's format is compatible with global integrative network's (GIN's) edgelist. When detected GRN network, it will automatically create new nodes and edges to fit the Petri Net design.
}
