% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{convert_graph_to_graphml}
\alias{convert_graph_to_graphml}
\title{Convert an igraph object into a list}
\usage{
convert_graph_to_graphml(g, cdglist, inputlist = "none")
}
\arguments{
\item{g}{The igraph object to be converted.}

\item{cdglist}{The list of coding genes' gene symbols.}

\item{inputlist}{The list of input tokens. Should be of two columns, the first is the text.show of the node. The second is the amount of tokens. Currently only works for RNA expressions.}
}
\value{
A list of two elements. The first one is the list of a header, nodes and edges, and an end, named "graphml". The second is "metainfo", which is a dataframe with three columns, id, textShow, and type.
}
\description{
Convert an igraph object into a list. If provided with an input list, then the RNA nodes which intersect with the input list will be given one extra input node and one input edge. The expression will be used as the weight of the edge.
}
