% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{convert_node_to_graphml}
\alias{convert_node_to_graphml}
\title{Convert a node into a list}
\usage{
convert_node_to_graphml(id.node.num, text.show, graphml, cdglist)
}
\arguments{
\item{id.node.num}{The number of a node id. A node id consists of "n" and a number, for example "n1", "n2"...}

\item{text.show}{The text.show used when creating a new node. If the node is a place node, then it will be the protein's name, RNA's name etc.. If the node is a transition node, it will be the symbol of the transition.}

\item{graphml}{The GraphML list.}

\item{cdglist}{The list of coding genes' gene symbols.}
}
\value{
A list of two elements. The first one is the text of a node, named "node_texts". The second is "metainfo", which is a dataframe with three columns, id, textShow, and type.
}
\description{
Generate a list of node text and the meta info. Meta info is useful for accelerating downstream conversion.
}
