% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{convert_nodeVec_to_graphml}
\alias{convert_nodeVec_to_graphml}
\title{Convert a vector of nodes into a list}
\usage{
convert_nodeVec_to_graphml(nodeVec, graphml, cdglist)
}
\arguments{
\item{nodeVec}{The vector of node names. NOT node ids.}

\item{graphml}{The GraphML list.}

\item{cdglist}{The list of coding genes' gene symbols.}
}
\value{
A list of two elements. The first one is the list of nodes and edges, named "graphml". The second is "metainfo", which is a dataframe with three columns, id, textShow, and type.
}
\description{
Generate a list of node texts and the meta info. Meta info is useful for accelerating downstream conversion.
}
