% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{create_neighbor_df}
\alias{create_neighbor_df}
\title{Create a dataframe containing the input gene/chemical list and their downstream neighbors}
\usage{
create_neighbor_df(tmp.gs, graph, num_cores = 6)
}
\arguments{
\item{tmp.gs}{A vector of input genes/chemicals.}

\item{graph}{The igraph object of Global Integrative Network (GIN).}

\item{num_cores}{The number of cores to use for parallel computation.}
}
\value{
A dataframe with two columns. The first column is the given genes. The second column is the downstream neighbors.
}
\description{
Create a dataframe of genes/chemicals in tmp.gs and their downstream neighbors by calling the function "neighbors" from package igraph. Has been paralleled.
}
