% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{decide_node_type}
\alias{decide_node_type}
\title{Distinguish the type of a given node}
\usage{
decide_node_type(node.text, cdglist.this)
}
\arguments{
\item{node.text}{The text.show of a node. Basically the symbols which will be shown in a GraphML editor for this node.}

\item{cdglist.this}{A list of coding genes' gene symbol.}
}
\value{
A character string of the node type. Can be one of "Catalysis", "Complex", "RNA", "Promoter", "Transcription", "Translation", "Binding", "Protein", and "Chemical".
}
\description{
Distinguish the type of a given node by its "text.show". This "text.show" was the same as the one used when creating this node.
}
