% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{examine_neighbor_df}
\alias{examine_neighbor_df}
\title{Check the types of the ending nodes, and split the df into two}
\usage{
examine_neighbor_df(df1, tmp.gs)
}
\arguments{
\item{df1}{A dataframe with two columns, starting nodes and ending nodes.}

\item{tmp.gs}{A vector of input gene/chemical list.}
}
\value{
A list of two dataframes. The first one contains the intermediate nodes in the second column. The second one contains genes/chemicals/complexes in the second column.
}
\description{
Check the types of the ending nodes, and split the df into two. One contains the intermeidate nodes, which should be further extended. The other one contains the products of reactions, which should stop the extension.
}
