% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{generate_ending}
\alias{generate_ending}
\title{Generate the ending of a graphml file}
\usage{
generate_ending(graphml, replace = FALSE)
}
\arguments{
\item{graphml}{A list object containing a header and other elements.}

\item{replace}{Logical. Whether to replace the last element in the list. Default is FALSE.}
}
\value{
A list containing a header, other elements and an end of a graphml file in mEPN format.
}
\description{
Generate the ending of a graphml file in mEPN schema.
}
