% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{generate_header}
\alias{generate_header}
\title{Generate the header of a graphml file}
\usage{
generate_header(graphml)
}
\arguments{
\item{graphml}{A list object with no content}
}
\value{
A list with only one element containing the texts of a header for a graphml file in mEPN format.
}
\description{
Generate a header for a graphml file in mEPN schema. The input object should be an empty list.
}
