% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{generate_metabolic_complex_node_collection}
\alias{generate_metabolic_complex_node_collection}
\title{Generate the collection of nodes in the structures of metabolic pathways or complex formations}
\usage{
generate_metabolic_complex_node_collection(input.gs, graph, cdglist)
}
\arguments{
\item{input.gs}{The input list of nodes. Could be the names of genes and chemicals}

\item{graph}{The igraph object of GIN. Compatible with GINv2's or GINv3's igraph object.}

\item{cdglist}{The list of coding genes' gene symbols.}
}
\value{
A vector of metabolic pathways' nodes inferred from GIN. Also contains nodes of complex formations.
}
\description{
Collect the nodes of mtabolic pathways or complex formations. Proteins usually function as enzymes in a metabolic reaction, therefore two proteins are separated by two nodes, an intermediate node and a chemical node. This is different from the signaling pathways, where proteins are separated by only one node, the intermediate node. This structure also fits the formation of complexes.
}
