% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{generate_node_collection}
\alias{generate_node_collection}
\title{Infer a node collection from Global Integrative Network based on an input list.}
\usage{
generate_node_collection(input.gs, graph, cdglist, num_cores = 6)
}
\arguments{
\item{input.gs}{The input list of nodes. Could be the names of genes and chemicals}

\item{graph}{The igraph object of GIN. Compatible with GINv2's or GINv3's igraph object.}

\item{cdglist}{The list of coding genes' gene symbols.}

\item{num_cores}{Numeric. The number of cores to use for parallel computation when generating the node collection.}
}
\value{
A vector of nodes inferred from GIN based on a give gene/chemical list.
}
\description{
This is a wrapper of two major functions which generate node collections of signaling pathways and metabolic pathways.
}
