% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{generate_PPrel_node_collection}
\alias{generate_PPrel_node_collection}
\title{Infer a node collection of signaling pathways from Global Integrative Network based on an input list.}
\usage{
generate_PPrel_node_collection(input.gs, graph, num_cores = 6)
}
\arguments{
\item{input.gs}{The input list of nodes. Could be the names of genes and chemicals}

\item{graph}{The igraph object of GIN. Compatible with GINv2's or GINv3's igraph object.}

\item{num_cores}{Numeric. The number of cores to use for parallel computation when generating the node collection.}
}
\value{
A vector of nodes of signaling pathways inferred from GIN based on a give gene/chemical list.
}
\description{
This is a wrapper of a core function which generate a node collection of signaling pathways. This function includes two steps, the first one is generating paths according to signaling pathways' structures. The second is to check these paths and collect the nodes.
}
