% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{generate_signaling_path}
\alias{generate_signaling_path}
\title{Collect the nodes of signaling pathways.}
\usage{
generate_signaling_path(input.gs, graph, num_cores = 6)
}
\arguments{
\item{input.gs}{The input list of genes and/or chemicals.}

\item{graph}{The igraph object of Global Integrative Network (GIN).}

\item{num_cores}{The number of cores to use for parallel computation.}
}
\value{
A vector of nodes of signaling pathways inferred from GIN.
}
\description{
The input gene/chemical list will be used as the seed to generate the all the possible paths. The paths will then be checked to obey the structure of signaling pathways.
}
