% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{get_node_by_path}
\alias{get_node_by_path}
\title{Collect the nodes of paths.}
\usage{
get_node_by_path(path.4check, graph, num_cores)
}
\arguments{
\item{path.4check}{A dataframe containing the paths to collect nodes. Two columns, starting nodes and ending nodes of the paths.}

\item{graph}{The igraph object of Global Integrative Network (GIN).}

\item{num_cores}{The number of cores to use for parallel computation.}
}
\value{
A vector of nodes of signaling pathways inferred from GIN.
}
\description{
Collect the nodes of paths in path.4check. Will call the function "get_node_by_path_epoch". Will add ".RNA" nodes if there are ".state" nodes (".state" nodes are gene promoter regions).
}
