% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{get_node_by_path_epoch}
\alias{get_node_by_path_epoch}
\title{An epoch of collecting the nodes of a path.}
\usage{
get_node_by_path_epoch(i, path.4check, graph)
}
\arguments{
\item{i}{The index of the path, i.e. the row index of the dataframe path.4check.}

\item{path.4check}{A dataframe containing the paths to collect nodes. Two columns, starting nodes and ending nodes of the paths.}

\item{graph}{The igraph object of Global Integrative Network (GIN).}
}
\value{
A vector of nodes of one signaling pathway inferred from GIN.
}
\description{
This is an epoch of collecting the nodes from a dataframe of path. This function will use the "all_shortest_paths" function from igraph package to get all the shortest paths from the starting node (column 1, row i) to the ending node (column 2, row i).
}
