% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{new_binding}
\alias{new_binding}
\title{Generate the texts of a transition node in "Binding" style of mEPN}
\usage{
new_binding(id, text.show = "B")
}
\arguments{
\item{id}{The id of the node. Should be in the format of "n" plus a number, for example "n1", "n2"...}

\item{text.show}{The text.show parameter which will be passed to function "basic_node". Default is "B", which is the symbol of a Binding node in mEPN style.}
}
\value{
A character string of a "Binding" node.
}
\description{
Generate the texts of a transition node in "Binding" style of mEPN. Usually just provide the node id. Change of the text.show will change the type of the node.
}
