% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{new_degenerate}
\alias{new_degenerate}
\title{Generate the texts of a transition node in "Degenerate" style of mEPN}
\usage{
new_degenerate(id, text.show = "/")
}
\arguments{
\item{id}{The id of the node. Should be in the format of "n" plus a number, for example "n1", "n2"...}

\item{text.show}{The text.show parameter which will be passed to function "basic_node". Default is "/", which is the symbol of a Degenerate node in mEPN style.}
}
\value{
A character string of a "Degenerate" node.
}
\description{
Generate the texts of a transition node in "Degenerate" style of mEPN. Usually just provide the node id. Change of the text.show will change the type of the node.
}
