% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{new_delayPlace}
\alias{new_delayPlace}
\title{Generate the texts of a place node in "Delayed place" style of mEPN}
\usage{
new_delayPlace(id)
}
\arguments{
\item{id}{The id of the node. Should be in the format of "n" plus a number, for example "n1", "n2"...}
}
\value{
A character string of a "Delayed place" node.
}
\description{
Generate the texts of a place node in "Delayed place" style of mEPN. This type of places are usually coupled with the "Delayed transition" nodes to mimic the delayed signals of stimulus.
}
