% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{new_delayTransit}
\alias{new_delayTransit}
\title{Generate the texts of a transition node in "Delayed transition" style of mEPN}
\usage{
new_delayTransit(id)
}
\arguments{
\item{id}{The id of the node. Should be in the format of "n" plus a number, for example "n1", "n2"...}
}
\value{
A character string of a "Delayed transition" node.
}
\description{
Generate the texts of a transition node in "Delayed transition" style of mEPN. This type of transitions are usually coupled with the "Delayed place" nodes to mimic the delayed signals of stimulus.
}
