% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{new_edge_id_num}
\alias{new_edge_id_num}
\title{Generate a number for a new edge}
\usage{
new_edge_id_num(id.edge.num, graphml)
}
\arguments{
\item{id.edge.num}{Any number.}

\item{graphml}{A graphml list.}
}
\value{
A numerical number, which is the number of existing edge plus 1.
}
\description{
Generate a number for a new edge. The input id.edge.num is not strictly required now. Can be given with any number.
}
