% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{new_edge_text}
\alias{new_edge_text}
\title{Generate the text of a new edge.}
\usage{
new_edge_text(id, source, target, type, weight = "")
}
\arguments{
\item{id}{An edge id which has not been used.}

\item{source}{The source node id.}

\item{target}{The target node id.}

\item{type}{The type of the edge. Should be one of "positive", "negative" and "input".}

\item{weight}{Numerical. The weight of an Input edge. Can be ignored when creating a simple edge.}
}
\value{
A character string of an edge.
}
\description{
Generate the text of a new edge based upon its source node id, target node id, type and weight (if it's an Input edge).
}
