% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{new_input_edge}
\alias{new_input_edge}
\title{Generate the texts of a weighted edge in mEPN style}
\usage{
new_input_edge(id, source, target, weight)
}
\arguments{
\item{id}{The id of the edge. Should be in the format of "e" plus a number, for example "e1", "e2"...}

\item{source}{The source node's id of this edge.}

\item{target}{The target node's id of this edge.}

\item{weight}{The weight of this edge.}
}
\value{
A character string of a weighted edge.
}
\description{
Generate the texts of a weighted edge in mEPN style. Currently this type of edge is only used for input tokens.
}
