% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{new_input_h}
\alias{new_input_h}
\title{Generate the texts of a place node in "Input" style of mEPN}
\usage{
new_input_h(id)
}
\arguments{
\item{id}{The id of the node. Should be in the format of "n" plus a number, for example "n1", "n2"...}
}
\value{
A character string of a "Input" node.
}
\description{
Generate the texts of a place node in "Input" style of mEPN. The input node will be a flat rectangle.
}
