% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{new_node_text}
\alias{new_node_text}
\title{Generate the text of a new node.}
\usage{
new_node_text(id, text.show, type)
}
\arguments{
\item{id}{A node id which has not been used.}

\item{text.show}{The text which is required for a node. If the node is a place node, then usually text.show is the name of the entity, for example protein name, gene name etc.. If the node is a transition node, then text.show will reflect the type of the transition.}
}
\value{
A character string which is the text of a node.
}
\description{
Generate the text of a new node based upon its type.
}
