% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{new_simple_edge}
\alias{new_simple_edge}
\title{Generate the texts of a simple edge with no weights in mEPN style}
\usage{
new_simple_edge(id, source, target, type = c("standard", "t_shape"))
}
\arguments{
\item{id}{The id of the edge. Should be in the format of "e" plus a number, for example "e1", "e2"...}

\item{source}{The source node's id of this edge.}

\item{target}{The target node's id of this edge.}

\item{type}{The type of this edge. Should be either "standard" (positive) or "t_shape" (negative).}
}
\value{
A character string of a simple edge.
}
\description{
Generate the texts of a simple edge in mEPN style.
}
