% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{randomSelect}
\alias{randomSelect}
\title{A simple function to generate a random integar}
\usage{
randomSelect(x)
}
\arguments{
\item{x}{The max integar.}
}
\value{
A random number of integar, possible max is "x".
}
\description{
This function calls "sample", and set size to 1.
}
