% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GINtoSPN.R
\name{rm_unrelatedNodes}
\alias{rm_unrelatedNodes}
\title{Remove unrelated nodes from the node vector of the sub-graph induced from Global Integrative Network (GIN)}
\usage{
rm_unrelatedNodes(input.gs, tmp.node, graph)
}
\arguments{
\item{input.gs}{The input list of nodes. Could be the names of genes and chemicals}

\item{tmp.node}{A vector of nodes which were extracted from GIN based on input.gs. Could be much larger than input.gs, since lots of intermediate nodes and participating nodes will be included.}

\item{graph}{The igraph object of GIN. Compatible with GINv2's or GINv3's igraph object.}
}
\value{
A vector of cleaned nodes.
}
\description{
Generate a list of node texts and the meta info. Meta info is useful for accelerating downstream conversion.
}
